Resolved implicit integer conversion warnings.

+++ Samples/JotFormTest/Classes/Controller/CreateQuestionViewController.m
@@ -66,8 +66,7 @@
         [SVProgressHUD dismiss];
         
         if ( result != nil ) {
-            
-            int responseCode = [[result objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[result objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 200 || responseCode == 206 ) {
                 
@@ -85,10 +84,9 @@
         [SVProgressHUD dismiss];
         
         if ( error != nil ) {
-            int responseCode = [[error objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[error objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 401 ) {
-                
                 NSString *errMsg = [NSString stringWithFormat:@"%@\nPlease check if your API Key's permission is 'Read Access' or 'Full Access'. You can create form with API key for 'Full Access'", [error objectForKey:@"message"]];
                 
                 UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:@"JotFormAPISample" message:errMsg delegate:self cancelButtonTitle:@"Ok" otherButtonTitles:nil, nil];
